chore: file rename TocGenerator->TocGeneratorExt
